README

+++ src/main/java/com/lennertsoffers/pokemon_city_api/model/dto/StatisticsDto.java
@@ -23,5 +23,5 @@ public record StatisticsDto(
         int moneyCollected,
         int incomePerMinute,
         int maxedCitizens,
-        int score
+        long score
 ) {}
